+++ test/spec/utils.spec.js
@@ -85,3 +85,4 @@ describe( 'return postfixed filenames', () => {
         } );
     } );
 } );
+